Move rapply to osf/utils/functional.py, change all imports
